generator model of dice
this is the first working commit of generator model of dice ORM function generator. Previously there was a compiler and it generated dice Go model functions from a TOML defined .dice files but that was something much depending on how we defined and use the models. The new approach generated functions inside the model files and leave the whole execution and maintenance of the generated models to the developer. sort of like scaffolding commonly used CRUD operations for that model